fix: tests are failing due to new chmod syscall

+++ sysaug/src/syscalls.rs
@@ -488,7 +488,7 @@ pub const RAW_SYSCALL_INFOS: [Option<SyscallInfo>; MAX_RAW_SYSCALL_INFOS] = {
         define_getdents_syscall!(libc::SYS_getdents, 32, iter, next);
 
         // New System call in Linux v6.6: fchmodat2
-        let libc__SYS_fchmodat2: usize = 452;
+        let libc__SYS_fchmodat2: i64 = 452;
         define_dirfd_fileperms_syscall!(libc__SYS_fchmodat2, 2, 0, PermType::Chmod, 2, iter, next);
         if let Some(val) = iter[next].as_mut() {
             val.flags = Some(3);
